Update register component
