+++ src/app/app-routing.module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { RegisterComponent } from './register/register.component';
-
-const routes: Routes = [{ path: 'register', component: RegisterComponent }];
+import { RegisterComponent } from './components/register/register.component';
 import { AuthComponent } from './components/auth/auth.component';
 
-const routes: Routes = [{ path: 'login', component: AuthComponent }];
+const routes: Routes = [
+  { path: 'login', component: AuthComponent },
+  { path: 'register', component: RegisterComponent },
+];
 
 @NgModule({
   imports: [RouterModule.forRoot(routes)],
