+++ src/app/components/register/register.component.html
@@ -1,0 +1,1 @@
+<div>Register works!</div>
