+++ src/environments/environment.ts
@@ -5,7 +5,7 @@
 export const environment = {
   production: false,
   // url for backend
-  BACKEND_URL: "http://localhost:8081"
+  BACKEND_URL: 'http://localhost:8080',
 };
 
 /*
